validacion a traves de rabbit done
